// wardn/frontend/lib/api/generated/model/pendingInvitationRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.0.1
 */
import type { PendingInvitationReadScopeType } from './pendingInvitationReadScopeType';

export interface PendingInvitationRead {
  expiresAt: string;
  id: string;
  organizationId: string;
  organizationName: string;
  role: string;
  scopeType: PendingInvitationReadScopeType;
  workspaceId: string | null;
  workspaceName: string | null;
}